Extract pages dir
